pyproject.toml
[project]
name = "friday-tony-stark-demo"
version = "0.2.0"
description = "J.A.R.V.I.S — Personal AI for Dushyant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # ── Core framework ───────────────────────────────────────
    "fastmcp",
    "httpx",
    "python-dotenv",

    # ── Voice pipeline ───────────────────────────────────────
    "livekit-agents[deepgram,groq,openai,sarvam,silero]>=1.5.1",
    "livekit-plugins-google>=1.5.1",

    # ── Web tools ────────────────────────────────────────────
    "duckduckgo-search>=6.0",

    # ── Desktop control ──────────────────────────────────────
    "pyautogui",
    "pygetwindow",
    "Pillow",

    # ── Database ─────────────────────────────────────────────
    "asyncpg",
    "redis[asyncio]",

    # ── Wake word ────────────────────────────────────────────
    "sounddevice",
    "numpy",
    "SpeechRecognition",
]

[project.scripts]
friday      = "server:main"
friday_voice = "agent_friday:dev"
jarvis_wake  = "wake:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py", "agent_friday.py", "wake.py"]
